Added friend requests from user in server, moved friend requests to their own service
