+++ Flycatcher.Services/Enumerations/CallbackTypes.cs
@@ -15,12 +15,13 @@ namespace Flycatcher.Services.Enumerations
 
         // Need a more comprehensive list of callback types
         ServerInvite,               // Server invite received or responded to from the recipient's side
-        ChannelMessageRecieved,     // Channel message received, so a new message in a channel, updates the message list in the channel
+        ChannelMessageEvent,        // Channel message received, edited or deleted, updates the message list in the channel
         FriendRequest,              // Friend request received or responded to from the recipient's side
         ServerPropertyUpdated,      // Updates the left side bar, so channel names, server names, added or removed channels etc.
-        ServerMemberUpdated,        // Updates the right side bar, so member names, roles, added or removed members etc.
+        ServerUserUpdated,          // Updates the right side bar, so member names, roles, added or removed members etc.
+        UserServerListUpdated,      // Updates the server list when a user joins or leaves a server
         FriendsListUpdated,         // Updates the friends list for both RX and TX, so added or removed friends, friend requests etc.
         ChannelDeleted,             // Channel deleted
-        ServerDeleted,              // Server deleted
+        ServerDeleted               // Server deleted
     }
 }
