+++ Flycatcher.Services/MessageService.cs
@@ -51,7 +51,7 @@ namespace Flycatcher.Services
             };
 
             await messageQueryableRepository.Create(message);
-            await callbackService.NotifyAsync(CallbackType.Channel, channelId);
+            await callbackService.NotifyAsync(CallbackType.ChannelMessageEvent, channelId);
         }
 
         public async Task<Result> DeleteMessage(int messageId)
@@ -66,7 +66,7 @@ namespace Flycatcher.Services
             var channelId = message.ChannelId;
 
             await messageQueryableRepository.Delete(message);
-            await callbackService.NotifyAsync(CallbackType.Channel, channelId);
+            await callbackService.NotifyAsync(CallbackType.ChannelMessageEvent, channelId);
 
             return new Result(true);
         }
