+++ Flycatcher/Components/_Imports.razor
@@ -11,6 +11,7 @@
 @using Flycatcher.Components.Widgets
 @using Flycatcher.Components.Widgets.HomeWidgets
 @using Flycatcher.Components.Widgets.ServerWidgets
+@using Flycatcher.Components.Widgets.ServerWidgets.ChannelWidgets
 @using MudBlazor
 @using MudBlazor.Services
 @using Flycatcher.Services
